Cleanup code, move command validation
