+++ pkg/devfile/validate/events.go
@@ -4,39 +4,37 @@ import (
 	"fmt"
 	"strings"
 
-	adapterCommon "github.com/openshift/odo/pkg/devfile/adapters/common"
-	"github.com/openshift/odo/pkg/devfile/parser/data"
+	"github.com/openshift/odo/pkg/devfile/parser/data/common"
 	"k8s.io/klog"
 )
 
 // validateEvents validates all the devfile events
-func validateEvents(data data.DevfileData) error {
+func validateEvents(events common.DevfileEvents, commands map[string]common.DevfileCommand, components []common.DevfileComponent) error {
 
 	eventErrors := ""
-	events := data.GetEvents()
 
 	switch {
 	case len(events.PreStart) > 0:
 		klog.V(2).Info("Validating preStart events")
-		if preStartErr := isEventValid(data, events.PreStart, "preStart"); preStartErr != nil {
+		if preStartErr := isEventValid(events.PreStart, "preStart", commands, components); preStartErr != nil {
 			eventErrors += fmt.Sprintf("\n%s", preStartErr.Error())
 		}
 		fallthrough
 	case len(events.PostStart) > 0:
 		klog.V(2).Info("Validating postStart events")
-		if postStartErr := isEventValid(data, events.PostStart, "postStart"); postStartErr != nil {
+		if postStartErr := isEventValid(events.PostStart, "postStart", commands, components); postStartErr != nil {
 			eventErrors += fmt.Sprintf("\n%s", postStartErr.Error())
 		}
 		fallthrough
 	case len(events.PreStop) > 0:
 		klog.V(2).Info("Validating preStop events")
-		if preStopErr := isEventValid(data, events.PreStop, "preStop"); preStopErr != nil {
+		if preStopErr := isEventValid(events.PreStop, "preStop", commands, components); preStopErr != nil {
 			eventErrors += fmt.Sprintf("\n%s", preStopErr.Error())
 		}
 		fallthrough
 	case len(events.PostStop) > 0:
 		klog.V(2).Info("Validating postStop events")
-		if postStopErr := isEventValid(data, events.PostStop, "postStop"); postStopErr != nil {
+		if postStopErr := isEventValid(events.PostStop, "postStop", commands, components); postStopErr != nil {
 			eventErrors += fmt.Sprintf("\n%s", postStopErr.Error())
 		}
 	}
@@ -52,10 +50,9 @@ func validateEvents(data data.DevfileData) error {
 // isEventValid checks if events belonging to a specific event type are valid:
 // 1. Event should map to a valid devfile command
 // 2. Event commands should be valid
-func isEventValid(data data.DevfileData, eventNames []string, eventType string) error {
+func isEventValid(eventNames []string, eventType string, commands map[string]common.DevfileCommand, components []common.DevfileComponent) error {
 	eventErrorMsg := make(map[string][]string)
 	eventErrors := ""
-	commands := data.GetCommands()
 
 	for _, eventName := range eventNames {
 		isEventPresent := false
@@ -65,7 +62,7 @@ func isEventValid(data data.DevfileData, eventNames []string, eventType string) 
 				isEventPresent = true
 
 				// Check if the devfile command is valid
-				err := adapterCommon.ValidateCommand(data, command)
+				err := validateCommand(command, commands, components)
 				if err != nil {
 					klog.V(2).Infof("command %s is not valid: %s", command.GetID(), err.Error())
 					eventErrorMsg[strings.ToLower(eventName)] = append(eventErrorMsg[strings.ToLower(eventName)], err.Error())
